python/requirements.txt
flask
Pillow
langchain
langchain-openai
python-dotenv